delete(#219): delete exist subscriber
